back-end fin

+++ controllers/funtions.js
@@ -1,11 +1,12 @@
 
 const People = require("../Models/people");
+const User = require('../Models/newUser');
 const asyncHandler = require('express-async-handler')
 
 
 
 const Read  = asyncHandler(async(req,res) =>{
-        const readpeople = await People.find({});
+        const readpeople = await People.find({ user:req.user.id});
         res.status(200).json(readpeople);
    
 });
@@ -13,19 +14,38 @@ const Read  = asyncHandler(async(req,res) =>{
 const Create = asyncHandler(async(req,res) =>{
 
     if(!req.body.text){
-        res.status(400).json({message: "Please Add text Field"})
-    }else{
-        const humans = await People.create(req.body);
-            res.status(201).json(People);
-    }
+        res.status(405)
+        throw new Error ("Please Add text Field");
+    };
+
+        const humans = await People.create({
+            text: req.body.text,
+            user: req.user.id,
+        });
+            res.status(200).json(humans);
+    
 }) 
  
 ///////////////////////////////////////////////////////////////////////////
 const Updates = asyncHandler(async(req,res)=>{
 
-        const updateperson = await People.findByIdAndUpdate(req.params.id);
+        const updateperson = await People.findById(req.params.id);
+      
+
         if(!updateperson){
-            res.status(404).json({message: "User Was Not found"})}
+            res.status(400)
+            throw new Error ("You are not authorized to update this user")
+        };
+
+        if(!req.user){
+            res.status(401)
+            throw new Error ("User Not Found");
+        }
+
+        if (updateperson.user.toString() !== req.user.id) {
+            res.status(401)
+            throw new Error('User not authorized')
+          }
 
         const updatepersonal = await People.findByIdAndUpdate(req.params.id , req.body , {
             new:true,
@@ -37,13 +57,40 @@ const Updates = asyncHandler(async(req,res)=>{
 
  ///////////////////////////////////////////////////////////////////////////
 const Deletes = asyncHandler(async(req,res)=>{
-   
-        const deleteperson = await People.findByIdAndDelete(req.params.id);
-        if(!deleteperson){
-            res.status(404).json({message: "User Was Not found"})
-        }
-        res.status(200).json({"deleted ID":req.params.id});
+
+    const deletepersons = await People.findById(req.params.id);
+      
+
+    if(!deletepersons){
+        res.status(400)
+        throw new Error ("User not found")
+    };
+    // Check for user
+    if(!req.user){
+        res.status(401)
+        throw new Error ("User Not Found");
+    }
+    // Make sure the logged in user matches the goal user
+    if (deletepersons.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+      }
+    
+      await People.findByIdAndDelete(req.params.id);
+    
+     res.status(200).json({"deleted ID":req.params.id});
      
 });
 
+///////////////////////////////////////////////////////////////////////////
+
+
 module.exports = {Read, Create, Updates , Deletes}
+
+
+// const deleteperson = await People.findByIdAndDelete(req.params.id);
+//         if(!deleteperson){
+//             res.status(404).json({message: "User Was Not found"})
+//         }
+//         res.status(200).json({"deleted ID":req.params.id});
+
